requirements.txt
altair
numpy
pandas
pydeck
streamlit
pytube
google-generativeai